Add `admin` files that don't have student data to repo
This change allows for people attempting to utilize this project to
follow the same steps as the people at UCLS. In essence this means it
gives "admins" the ability to pickle and then distribute files to
students. See the README.md file for more information about how to
actually do this.